[pubsub] Return NULL value for PubSubProduceResult upon send failure

+++ clients/venice-push-job/src/test/java/com/linkedin/venice/hadoop/input/kafka/KafkaInputRecordReaderTest.java
@@ -72,7 +72,8 @@ public class KafkaInputRecordReaderTest {
       put.putValue = ByteBuffer.wrap(valueBytes);
       put.replicationMetadataPayload = ByteBuffer.allocate(0);
       messageEnvelope.payloadUnion = put;
-      consumerRecordList.add(new ImmutablePubSubMessage<>(kafkaKey, messageEnvelope, pubSubTopicPartition, i, -1, -1));
+      consumerRecordList
+          .add(new ImmutablePubSubMessage<>(kafkaKey, messageEnvelope, pubSubTopicPartition, (long) i, -1, -1));
     }
 
     Map<PubSubTopicPartition, List<PubSubMessage<KafkaKey, KafkaMessageEnvelope, Long>>> recordsMap = new HashMap<>();
